Use va-link-action component
